backend/requirements.txt
astroid==2.4.2
autopep8==1.5.3
beautifulsoup4==4.9.1
certifi==2020.6.20
chardet==3.0.4
colorama==0.4.3
Django==2.2
django-cors-headers==3.4.0
django-filter==2.3.0
django-haystack==2.8.1
django-js-asset==1.2.2
django-mptt==0.11.0
django-sortedm2m==3.0.2
djangorestframework==3.10.3
drf-haystack==1.8.6
elasticsearch==2.4.0
idna==2.10
importlib-metadata==1.7.0
isort==4.3.21
itsdangerous==1.1.0
lazy-object-proxy==1.4.3
Markdown==3.2.2
mccabe==0.6.1
pycodestyle==2.6.0
pylint==2.5.3
python-dateutil==2.8.1
pytz==2020.1
qiniu==7.2.8
requests==2.24.0
six==1.15.0
soupsieve==2.0.1
sqlparse==0.3.1
toml==0.10.1
typed-ast==1.4.1
urllib3==1.22
wincertstore==0.2
wrapt==1.12.1
zipp==3.1.0
mysqlclient==1.3.14
gunicorn==20.0.4